improve ship embarking
update to use the tillerman and gump buttons to perform embarking. this way we can check gump id to see if it is a boat menu that we can use, or search for the next tillerman.

fixes issues where you can embark on another player's boat when hunting in groups